+++ carno-livechat.php
@@ -2,8 +2,9 @@
 /**
  * Plugin Name: Carno LiveChat
  * Description: One-way admin broadcast chat system presented as a chat UI.
- * Version:     1.3.0
- * Author:      Carno
+ * Version:     1.4.0
+ * Author:      سپهر علیمحمدی
+ * Author-URI:  https://sepehralimohammadi.com
  * Text Domain: carno-livechat
  * Domain Path: /languages
  */
@@ -12,7 +13,7 @@ if ( ! defined( 'ABSPATH' ) ) {
     exit;
 }
 
-define( 'CARNO_LIVECHAT_VERSION',  '1.3.0' );
+define( 'CARNO_LIVECHAT_VERSION',  '1.4.0' );
 define( 'CARNO_LIVECHAT_FILE',     __FILE__ );
 define( 'CARNO_LIVECHAT_PATH',     plugin_dir_path( __FILE__ ) );
 define( 'CARNO_LIVECHAT_URL',      plugin_dir_url( __FILE__ ) );